Changed WordExtensions to WordMeasurer
